pyproject.toml
[project]
name = "weekly-marketing-report"
version = "0.1.0"
description = "Weekly site intelligence report for claudecodeformarketers.com"
requires-python = ">=3.11"
dependencies = [
    "python-frontmatter>=1.1.0",
    "boto3>=1.34.0",
    "google-api-python-client>=2.140.0",
    "google-auth>=2.34.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1",
    "resend>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
